refactor: ViewModel의 Android Framework 의존성 제거 및 상세 화면 컴포넌트 리팩터링
- ViewModel에서 `Uri`, `android.util.Patterns` 의존성을 제거하고 `String` 및 자체 정규식(`EMAIL_ADDRESS_REGEX`)을 사용하도록 변경
- `ReceiverAfterNoteMainScreen`을 `MemorialReceivedDetailScreen`으로 이름을 변경하고 수신자 측 상세 화면 프로토타입으로 재정의
- `AddSongError` sealed interface를 추가하여 에러 메시지 결정(String resolve) 로직을 UI 레이어로 위임
- `OnboardingProfileEntry`에서 ViewModel의 `String` 데이터를 UI용 `Uri`로 변환하는 매핑 로직 추가
- 불필요한 네비게이션 임포트 제거 및 컴포저블 KDoc 설명 보완
